requirements.txt

Django~=2.1
python-dateutil